+++ history.md
@@ -1,6 +1,7 @@
 # 2023-01-23
 
 - Make `vscode.open` command work with file uri.
+- Cancel option for `workspace.registerExprKeymap()`.
 
 # 2023-01-22